requirements.txt
fastapi
uvicorn
azure-storage-blob
python-multipart
python-dotenv
azure-identity
pydantic[email]
azure-data-tables
python-jose[cryptography]
requests
